+++ src/Enum.hs
@@ -1,5 +1,9 @@
 module Enum
   ( Enum(..)
+  , toEnumMay
+  , succMay
+  , predMay
   ) where
 
 import GHC.Enum
+import Safe     (predMay, succMay, toEnumMay)