+++ src/List.hs
@@ -30,6 +30,8 @@ module List
   , filter
   , findIndex
   , findIndices
+  , foldl1May'
+  , foldr1May
   , genericDrop
   , genericIndex
   , genericLength
@@ -42,6 +44,8 @@ module List
   , groupSort
   , groupSortBy
   , groupSortOn
+  , headMay
+  , initMay
   , inits
   , insert
   , insertBy
@@ -57,8 +61,13 @@ module List
 #if MIN_VERSION_base(4,11,0)
   , iterate'
 #endif
+  , lastMay
   , lookup
   , map
+  , maximumByMay
+  , maximumMay
+  , minimumByMay
+  , minimumMay
   , nub
   , nubBy
   , nubOn
@@ -93,6 +102,7 @@ module List
   , stripPrefix
   , stripSuffix
   , subsequences
+  , tailMay
   , tails
   , take
   , takeEnd
@@ -123,6 +133,12 @@ module List
   , zipWith5
   , zipWith6
   , zipWith7
+    -- ** Diff algorithm
+  , Diff(..)
+  , getDiff
+  , getDiffBy
+  , getGroupedDiff
+  , getGroupedDiffBy
     -- ** String
   , String
   , words
@@ -142,20 +158,25 @@ module List
   , suffixed
   ) where
 
-import Data.List (break, cycle, delete, deleteBy, deleteFirstsBy, drop,
-                  dropWhile, dropWhileEnd, elemIndex, elemIndices, filter,
-                  findIndex, findIndices, genericDrop, genericIndex,
-                  genericLength, genericReplicate, genericSplitAt, genericTake,
-                  group, groupBy, inits, insert, insertBy, intercalate,
-                  intersect, intersectBy, intersperse, isInfixOf, isPrefixOf,
-                  isSubsequenceOf, isSuffixOf, iterate, lines, lookup, map, nub,
-                  nubBy, partition, permutations, repeat, replicate, reverse,
-                  scanl, scanl', scanl1, scanr, scanr1, sort, sortBy, sortOn,
-                  span, splitAt, stripPrefix, subsequences, tails, take,
-                  takeWhile, transpose, uncons, unfoldr, union, unionBy,
-                  unlines, unwords, unzip, unzip3, unzip4, unzip5, unzip6,
-                  unzip7, words, zip, zip3, zip4, zip5, zip6, zip7, zipWith,
-                  zipWith3, zipWith4, zipWith5, zipWith6, zipWith7, (++), (\\))
+import Data.Algorithm.Diff (Diff(Both, First, Second), getDiff, getDiffBy,
+                            getGroupedDiff, getGroupedDiffBy)
+import Data.List           (break, cycle, delete, deleteBy, deleteFirstsBy,
+                            drop, dropWhile, dropWhileEnd, elemIndex,
+                            elemIndices, filter, findIndex, findIndices,
+                            genericDrop, genericIndex, genericLength,
+                            genericReplicate, genericSplitAt, genericTake,
+                            group, groupBy, inits, insert, insertBy,
+                            intercalate, intersect, intersectBy, intersperse,
+                            isInfixOf, isPrefixOf, isSubsequenceOf, isSuffixOf,
+                            iterate, lines, lookup, map, nub, nubBy, partition,
+                            permutations, repeat, replicate, reverse, scanl,
+                            scanl', scanl1, scanr, scanr1, sort, sortBy, sortOn,
+                            span, splitAt, stripPrefix, subsequences, tails,
+                            take, takeWhile, transpose, uncons, unfoldr, union,
+                            unionBy, unlines, unwords, unzip, unzip3, unzip4,
+                            unzip5, unzip6, unzip7, words, zip, zip3, zip4,
+                            zip5, zip6, zip7, zipWith, zipWith3, zipWith4,
+                            zipWith5, zipWith6, zipWith7, (++), (\\))
 #if MIN_VERSION_base(4,11,0)
 import Data.List (iterate')
 #endif
@@ -171,3 +192,6 @@ import Data.List.Split (chop, divvy, endBy, splitOn)
 import Data.String     (IsString(fromString), String)
 import GHC.Show        (showString)
 import Numeric         (lexDigits)
+import Safe            (foldl1May', foldr1May, headMay, initMay, lastMay,
+                        maximumByMay, maximumMay, minimumByMay, minimumMay,
+                        tailMay)