+++ src/Num/Integer/Partial.hs
@@ -1,0 +1,6 @@
+module Num.Integer.Partial
+  ( digits
+  , undigits
+  ) where
+
+import Data.FastDigits (digits, undigits)