+++ src/Num/Natural/Partial.hs
@@ -1,0 +1,5 @@
+module Num.Natural.Partial
+  ( digitsUnsigned
+  ) where
+
+import Data.FastDigits (digitsUnsigned)